+++ backend/src/main/java/com/back/back9/domain/tradeLog/dto/ProfitAnalysisDto.java
@@ -3,7 +3,7 @@ package com.back.back9.domain.tradeLog.dto;
 import java.math.BigDecimal;
 
 public record ProfitAnalysisDto(
-        Long coinId,
+        int coinId,
         BigDecimal totalQuantity,
         BigDecimal averageBuyPrice,
         BigDecimal realizedProfitRate
