+++ backend/src/main/java/com/back/back9/domain/tradeLog/dto/ProfitRateResponse.java
@@ -3,6 +3,6 @@ package com.back.back9.domain.tradeLog.dto;
 import java.util.List;
 
 public record ProfitRateResponse(
-        Long userId,
+        int userId,
         List<ProfitAnalysisDto> coinAnalytics
 ){}
